+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [unreleased]
 
+* Implemented a viewer + editor for all the cached values (authentication tokens, database names, etc.)
+* Improved authentication logic to better handle multi-account scenarios
 
 ## [0.1.1]